+++ src/index.html
@@ -24,12 +24,12 @@
                 width=50px>
             <h2 class="questionbox__headline">Frage 1</h2>
             <p class="questionbox__question">Welches Organ „ernährt“ sich überwiegend von Kohlenhydraten?</p>
-            <p class="questionbox__answer" id="answer1">Die Anwort ist das Gehirn.</p>
+            <p class="questionbox__answer" id="answer1">Das Gehirn.</p>
             <button class="questionbox__button" id="button1">Antwort anzeigen</button>
 
 
             <ul class="questionbox__tags">
-                <li class="questionbox__tags__1">Das Gehirn</li>
+                <li class="questionbox__tags__1 right">Das Gehirn</li>
                 <li class="questionbox__tags__2">Der Darm</li>
                 <li class="questionbox__tags__3">Die Haut</li>
                 <li class="questionbox__tags__4">Die Leber</li>
@@ -41,15 +41,15 @@
                 width=50px>
             <h2 class="questionbox__headline">Frage 2</h2>
             <p class="questionbox__question">Welcher Mangel kann bei rein vegetarischer Ernährung auftreten?</p>
-            <p class="questionbox__answer" id="answer2">Die Anwort ist Eisenmangel.</p>
+            <p class="questionbox__answer" id="answer2">Der Eisenmangel.</p>
             <button class="questionbox__button" id="button2">Antwort anzeigen</button>
 
 
             <ul class="questionbox__tags">
                 <li class="questionbox__tags__1">Eiweißmangel</li>
-                <li class="questionbox__tags__1">Vitamin-D-Mangel</li>
-                <li class="questionbox__tags__1">Eisenmangel</li>
-                <li class="questionbox__tags__1">Zinkmangel</li>
+                <li class="questionbox__tags__2">Vitamin-D-Mangel</li>
+                <li class="questionbox__tags__3 right">Eisenmangel</li>
+                <li class="questionbox__tags__4">Zinkmangel</li>
             </ul>
         </section>
 
@@ -58,14 +58,14 @@
                 width=50px>
             <h2 class="questionbox__headline">Frage 3</h2>
             <p class="questionbox__question">Welches Gemüse enthält pro 100 Gramm die meisten Kohlenhydrate?</p>
-            <p class="questionbox__answer" id="answer3">Die Anwort sind Erbsen.</p>
+            <p class="questionbox__answer" id="answer3">Erbsen.</p>
             <button class="questionbox__button" id="button3">Antwort anzeigen</button>
 
             <ul class="questionbox__tags">
                 <li class="questionbox__tags__1">Kürbis</li>
-                <li class="questionbox__tags__1">Erbsen</li>
-                <li class="questionbox__tags__1">Karotten</li>
-                <li class="questionbox__tags__1">Sellerie</li>
+                <li class="questionbox__tags__2 right">Erbsen</li>
+                <li class="questionbox__tags__3">Karotten</li>
+                <li class="questionbox__tags__4">Sellerie</li>
             </ul>
         </section>
 
@@ -74,14 +74,14 @@
                 width=50px>
             <h2 class="questionbox__headline">Frage 4</h2>
             <p class="questionbox__question">Wie viel Gramm Zucker empfiehlt die WHO als Höchstmenge pro Tag?</p>
-            <p class="questionbox__answer" id="answer4">Die Anwort ist 25 Gramm.</p>
+            <p class="questionbox__answer" id="answer4">25 Gramm.</p>
             <button class="questionbox__button" id="button4">Antwort anzeigen</button>
 
             <ul class="questionbox__tags">
                 <li class="questionbox__tags__1">50 Gramm</li>
-                <li class="questionbox__tags__1">40 Gramm</li>
-                <li class="questionbox__tags__1">25 Gramm</li>
-                <li class="questionbox__tags__1">15 Gramm</li>
+                <li class="questionbox__tags__2">40 Gramm</li>
+                <li class="questionbox__tags__3 right">25 Gramm</li>
+                <li class="questionbox__tags__4">15 Gramm</li>
             </ul>
         </section>
     </main>
